requirements.txt
dj-database-url==0.4.2
Django==2.0.1
djangorestframework==3.7.7
Pillow==5.0.0
pytz==2017.3
whitenoise==4.0b4
